64-bit Linux compilation issue : fixes the INT64 / UINT64 type mismatches and missing prototype for memset().

+++ FreeImage/Source/LibRawLite/libraw/libraw_types.h
@@ -73,8 +73,9 @@ extern "C" {
 typedef __int64 INT64;
 typedef unsigned __int64 UINT64;
 #else
-typedef long long INT64;
-typedef unsigned long long UINT64;
+#include <stdint.h>
+typedef int64_t INT64;
+typedef uint64_t UINT64;
 #endif
 
 typedef unsigned char uchar;
